稍微修改了一下结构

+++ BiliAPI/BiliDynamic/DynamicEntity/BiliH5Dynamic/BiliH5DynamicSketch.cs
@@ -6,7 +6,7 @@
         public string desc_text { get; set; }
         public string cover_url { get; set; }
         public string target_url { get; set; }
-        public int sketch_id { get; set; }
+        public long sketch_id { get; set; }
         public int biz_type { get; set; }
         public BiliH5DynamicTagsItem[] tags { get; set; }
     }
